mixed: renaming aesthetics
