+++ opm/simulators/linalg/setupPropertyTree.cpp
@@ -253,16 +253,6 @@ setupPropertyTree(FlowLinearSolverParameters p, // Note: copying the parameters 
     }
 
     // mixed-precision ILU0
-    if (conf == "haugen-ilu0") {
-        return setupHaugenILU(conf, p);
-    }
-
-    // mixed-precision ILU0
-    if (conf == "haugen-dilu") {
-        return setupHaugenDILU(conf, p);
-    }
-
-    // mixed-precision ILU0
     if (conf == "mixed-ilu0") {
         return setupMixedILU(conf, p);
     }
@@ -270,6 +260,16 @@ setupPropertyTree(FlowLinearSolverParameters p, // Note: copying the parameters 
     // mixed-precision ILU0
     if (conf == "mixed-dilu") {
         return setupMixedDILU(conf, p);
+    }
+
+    // mixed-precision ILU0
+    if (conf == "legacy-mixed-ilu0") {
+        return setupLegacyMixedILU(conf, p);
+    }
+
+    // mixed-precision ILU0
+    if (conf == "legacy-mixed-dilu") {
+        return setupLegacyMixedDILU(conf, p);
     }
 
     if (conf == "dilu") {
@@ -435,33 +435,6 @@ setupILU([[maybe_unused]] const std::string& conf, const FlowLinearSolverParamet
 }
 
 PropertyTree
-setupHaugenILU([[maybe_unused]] const std::string& conf, const FlowLinearSolverParameters& p)
-{
-    using namespace std::string_literals;
-    PropertyTree prm;
-    prm.put("tol", p.linear_solver_reduction_);
-    prm.put("maxiter", p.linear_solver_maxiter_);
-    prm.put("verbosity", p.linear_solver_verbosity_);
-    prm.put("solver", "mixed-precision"s);
-    prm.put("preconditioner.type", "haugen-ilu0"s);
-    return prm;
-}
-
-PropertyTree
-setupHaugenDILU([[maybe_unused]] const std::string& conf, const FlowLinearSolverParameters& p)
-{
-    using namespace std::string_literals;
-    PropertyTree prm;
-    prm.put("tol", p.linear_solver_reduction_);
-    prm.put("maxiter", p.linear_solver_maxiter_);
-    prm.put("verbosity", p.linear_solver_verbosity_);
-    prm.put("solver", "mixed-precision"s);
-    prm.put("preconditioner.type", "haugen-dilu"s);
-    return prm;
-}
-
-
-PropertyTree
 setupMixedILU([[maybe_unused]] const std::string& conf, const FlowLinearSolverParameters& p)
 {
     using namespace std::string_literals;
@@ -469,7 +442,7 @@ setupMixedILU([[maybe_unused]] const std::string& conf, const FlowLinearSolverPa
     prm.put("tol", p.linear_solver_reduction_);
     prm.put("maxiter", p.linear_solver_maxiter_);
     prm.put("verbosity", p.linear_solver_verbosity_);
-    prm.put("solver", "mixed-bicgstab"s);
+    prm.put("solver", "mixed-precision"s);
     prm.put("preconditioner.type", "mixed-ilu0"s);
     return prm;
 }
@@ -482,8 +455,35 @@ setupMixedDILU([[maybe_unused]] const std::string& conf, const FlowLinearSolverP
     prm.put("tol", p.linear_solver_reduction_);
     prm.put("maxiter", p.linear_solver_maxiter_);
     prm.put("verbosity", p.linear_solver_verbosity_);
-    prm.put("solver", "mixed-bicgstab"s);
+    prm.put("solver", "mixed-precision"s);
     prm.put("preconditioner.type", "mixed-dilu"s);
+    return prm;
+}
+
+
+PropertyTree
+setupLegacyMixedILU([[maybe_unused]] const std::string& conf, const FlowLinearSolverParameters& p)
+{
+    using namespace std::string_literals;
+    PropertyTree prm;
+    prm.put("tol", p.linear_solver_reduction_);
+    prm.put("maxiter", p.linear_solver_maxiter_);
+    prm.put("verbosity", p.linear_solver_verbosity_);
+    prm.put("solver", "mixed-bicgstab"s);
+    prm.put("preconditioner.type", "legacy-mixed-ilu0"s);
+    return prm;
+}
+
+PropertyTree
+setupLegacyMixedDILU([[maybe_unused]] const std::string& conf, const FlowLinearSolverParameters& p)
+{
+    using namespace std::string_literals;
+    PropertyTree prm;
+    prm.put("tol", p.linear_solver_reduction_);
+    prm.put("maxiter", p.linear_solver_maxiter_);
+    prm.put("verbosity", p.linear_solver_verbosity_);
+    prm.put("solver", "mixed-bicgstab"s);
+    prm.put("preconditioner.type", "legacy-mixed-dilu"s);
     return prm;
 }
