+++ opm/simulators/linalg/setupPropertyTree.hpp
@@ -48,11 +48,11 @@ void validateSystemCPRTree(const PropertyTree& prm);
 void checkSystemCPRMatrixAddWell(bool matrixAddWellContributions);
 PropertyTree setupAMG(const std::string& conf, const FlowLinearSolverParameters& p);
 PropertyTree setupILU(const std::string& conf, const FlowLinearSolverParameters& p);
+PropertyTree setupLegacyMixedILU(const std::string& conf, const FlowLinearSolverParameters& p);
 PropertyTree setupMixedILU(const std::string& conf, const FlowLinearSolverParameters& p);
-PropertyTree setupHaugenILU(const std::string& conf, const FlowLinearSolverParameters& p);
 PropertyTree setupDILU(const std::string& conf, const FlowLinearSolverParameters& p);
+PropertyTree setupLegacyMixedDILU(const std::string& conf, const FlowLinearSolverParameters& p);
 PropertyTree setupMixedDILU(const std::string& conf, const FlowLinearSolverParameters& p);
-PropertyTree setupHaugenDILU(const std::string& conf, const FlowLinearSolverParameters& p);
 PropertyTree setupUMFPack(const std::string& conf, const FlowLinearSolverParameters& p);
 
 } // namespace Opm
